pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-review-env"
version = "1.0.0"
description = "OpenEnv-compliant RL environment for software code review"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}

dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-ulid>=2.0",
    "scipy>=1.10",
    "numpy>=1.24",
    "openai>=1.0",
]

[project.scripts]
code-review-env = "server.app:main"
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
research = [
    "torch>=2.0",
    "sentence-transformers>=2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "env*",
    "graders*",
    "tasks*",
    "benchmark*",
    "analysis*",
    "world_model*",
    "server*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
